Improve support for location hyperlinks
If multiple locations are detected for a region, try to create a single
hyperlink which will show additional UI that allows the user to select
from a list of locations; failing that, fallback to the old behavior
of creating a hyperlink for each of the detected locations.

DefinitionHyperlinkDetector and TypeDefinitionHyperlinkDetector now use
ShowSearchResultHyperlink to show multiple detected locations in the
search result view. To enable this, the hyperlink detector's context
should be able to provide a DocumentService; for an example, see
TypeScriptSourceViewerConfiguration#getHyperlinkDetectors.