Add support for multiple providers to `ContentAssistProcessor`
Multiple completion providers can be registered for a language. In that
case providers are sorted by their score and groups of equal score are
sequentially asked for completion items. The process stops when one or
many providers of a group return a result. A failing provider will not
fail the whole operation.

Similarly, multiple signature help providers can be registered for a
language. In that case providers are sorted by their score and called
sequentially until a provider returns a valid result.