Support multiple folding range providers in LX4E
Multiple folding range providers can be registered for a language. In
that case providers are asked in parallel and the results are merged. If
multiple folding ranges start at the same position, only the range of
the first registered provider is used. If a folding range overlaps with
another range that has a smaller start position, it is also ignored. A
failing provider will not cause a failure of the whole operation.